+++ src/conversations/index.ts
@@ -2,8 +2,16 @@ import type { App } from 'vue';
 import ConversationsComponent from './Conversations.vue';
 import Creation from './Creation.vue';
 
-export type { ConversationsProps, Conversation, CreationProps, ShortcutKeys } from './interface';
-export { ShortcutKeyCode } from './interface';
+export type {
+  ConversationsProps,
+  Conversation,
+  ConversationItemType,
+  DividerItemType,
+  ConversationsItems,
+  CreationProps,
+  ShortcutKeys,
+} from './interface';
+export { ShortcutKeyCode, isDividerItem } from './interface';
 
 type Compounded = typeof ConversationsComponent & {
   Creation: typeof Creation;